section 8 (partial): split extents code from hfs_fsck
src/fs/hfs_fsck.rs (6588 lines) -> src/fs/hfs_fsck/{mod,extents}.rs:
- mod.rs: 5279 lines (catalog + MDB + B-tree + bitmap phases)
- extents.rs: 1337 lines (extents-overflow B-tree checks + repairs + tests)

The 17 *_extents_* functions plus their dedicated test block move to
extents.rs. Helpers HfsFsckCode, HFS_MAX_NRECS, hfs_issue,
extract_child_pointers, record_key bumped to pub(super) so the new
submodule can reach them. All 61 fsck tests still pass.

Further per-phase splits (mdb, catalog, btree, bitmap) deferred — the
remaining mod.rs has tighter cross-references between those phases.
